refactor: dedupe into shared modules, split book calendar, group files into folders
