+++ locale/hu.po
